removed content-type for multipart/form-data
When uploading files via Ajax and FormData, you must not set the `Content-Type: multipart/form-data` header. This will cause the browser to really send a header Content-Type: multipart/form-data which is wrong. If you do not set this header explicitely it will set the header to something like Content-Type: multipart/form-data; boundary=----WebKitFormBoundaryhwk47FtsDeemWl3V. So basically the boundary is missing which makes it impossible for the server side to parse the body.